+++ app/src/main/java/nextstep/github/data/remote/mapper/RepositoryResponseToDataMapper.kt
@@ -1,11 +1,11 @@
 package nextstep.github.data.remote.mapper
 
-import nextstep.github.data.model.RepositoryModel
+import nextstep.github.data.dto.RepositoryDto
 import nextstep.github.data.remote.response.RepositoryResponse
 
-fun List<RepositoryResponse>.toDataList(): List<RepositoryModel> = map { it.toData() }
+fun List<RepositoryResponse>.toDataList(): List<RepositoryDto> = map { it.toData() }
 
-fun RepositoryResponse.toData() = RepositoryModel(
+fun RepositoryResponse.toData() = RepositoryDto(
     id = id,
     fullName = fullName ?: "",
     description = description ?: "",
