+++ app/src/main/java/nextstep/github/data/repository/FakeGithubRepository.kt
@@ -1,13 +1,13 @@
 package nextstep.github.data.repository
 
-import nextstep.github.data.model.RepositoryModel
+import nextstep.github.data.dto.RepositoryDto
 
 class FakeGithubRepository : GithubRepository {
 
     private var shouldFail = false
-    private var fakeData: List<RepositoryModel> = emptyList()
+    private var fakeData: List<RepositoryDto> = emptyList()
 
-    fun setFakeData(data: List<RepositoryModel>) {
+    fun setFakeData(data: List<RepositoryDto>) {
         shouldFail = false
         fakeData = data
     }
@@ -16,7 +16,7 @@ class FakeGithubRepository : GithubRepository {
         shouldFail = isFail
     }
 
-    override suspend fun getRepositories(organization: String): Result<List<RepositoryModel>> {
+    override suspend fun getRepositories(organization: String): Result<List<RepositoryDto>> {
         return if (shouldFail) {
             Result.failure(Exception("Fake network error"))
         } else {
