+++ app/src/main/java/nextstep/github/data/repository/GithubRepository.kt
@@ -1,18 +1,18 @@
 package nextstep.github.data.repository
 
-import nextstep.github.data.model.RepositoryModel
+import nextstep.github.data.dto.RepositoryDto
 import nextstep.github.data.remote.api.GithubService
 import nextstep.github.data.remote.mapper.toDataList
 
 interface GithubRepository {
-    suspend fun getRepositories(organization: String): Result<List<RepositoryModel>>
+    suspend fun getRepositories(organization: String): Result<List<RepositoryDto>>
 }
 
 class GithubRepositoryImpl(
     private val githubService: GithubService,
 ) : GithubRepository {
 
-    override suspend fun getRepositories(organization: String): Result<List<RepositoryModel>> {
+    override suspend fun getRepositories(organization: String): Result<List<RepositoryDto>> {
         return runCatching {
             githubService
                 .getRepositories(organization = organization)
